requirements.txt
PyMuPDF==1.27.2
opencv-python-headless==4.11.0.86
numpy==1.26.4
pillow==11.1.0
pandas==2.3.2
openpyxl==3.1.5
pytesseract==0.3.13
paddleocr==3.4.0
paddlepaddle==3.3.0
